Extracted some duplicate offline recovery code Should make it easier to manage rather than having to change it everywhere each time
