+++ src/main/java/com/buuz135/togetherforever/TogetherForever.java
@@ -48,7 +48,7 @@ public class TogetherForever {
 
     public static Logger LOGGER;
 
-    //Add a comand to force sync players
+    //Add a command to force sync players
     //Sync other player stuff when player joins
     /**
      * This is the instance of your mod as created by Forge. It will never be null.
