+++ src/main/java/com/buuz135/togetherforever/action/recovery/GameStageOfflineRecovery.java
@@ -1,11 +1,7 @@
 package com.buuz135.togetherforever.action.recovery;
 
 import com.buuz135.togetherforever.action.GameStagesEventSyncAction;
-import com.buuz135.togetherforever.api.IOfflineSyncRecovery;
 import com.buuz135.togetherforever.api.IPlayerInformation;
-import com.buuz135.togetherforever.api.data.TogetherRegistries;
-import com.google.common.collect.ArrayListMultimap;
-import com.google.common.collect.ListMultimap;
 import net.darkhax.gamestages.GameStageHelper;
 import net.minecraft.nbt.NBTTagCompound;
 
@@ -13,24 +9,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-public class GameStageOfflineRecovery implements IOfflineSyncRecovery {
-
-    private ListMultimap<IPlayerInformation, NBTTagCompound> offlineRecoveries;
-
+public class GameStageOfflineRecovery extends AbstractOfflineRecovery {
     public GameStageOfflineRecovery() {
-        this.offlineRecoveries = ArrayListMultimap.create();
-    }
-
-    @Override
-    public void storeMissingPlayers(List<IPlayerInformation> playerInformations, NBTTagCompound store) {
-        for (IPlayerInformation playerInformation : playerInformations) {
-            storeMissingPlayer(playerInformation, store);
-        }
-    }
-
-    @Override
-    public void storeMissingPlayer(IPlayerInformation playerInformation, NBTTagCompound store) {
-        offlineRecoveries.put(playerInformation, store);
+        super();
     }
 
     @Override
@@ -47,46 +28,6 @@ public class GameStageOfflineRecovery implements IOfflineSyncRecovery {
         }
         for (Map.Entry<IPlayerInformation, NBTTagCompound> entry : removeList) {
             offlineRecoveries.remove(entry.getKey(), entry.getValue());
-        }
-    }
-
-    @Override
-    public NBTTagCompound writeToNBT() {
-        NBTTagCompound tagCompound = new NBTTagCompound();
-        for (IPlayerInformation playerInformation : offlineRecoveries.keySet()) {
-            String uuid = playerInformation.getUUID().toString();
-            NBTTagCompound recovery = new NBTTagCompound();
-            recovery.setTag("ID", playerInformation.getNBTTag());
-            recovery.setString("PlayerID", TogetherRegistries.getPlayerInformationID(playerInformation.getClass()));
-            int id = 0;
-            for (NBTTagCompound compound : offlineRecoveries.get(playerInformation)) {
-                recovery.setTag(Integer.toString(id), compound);
-                ++id;
-            }
-            tagCompound.setTag(uuid, recovery);
-        }
-        return tagCompound;
-    }
-
-    @Override
-    public void readFromNBT(NBTTagCompound compound) {
-        offlineRecoveries.clear();
-        for (String uuid : compound.getKeySet()) {
-            NBTTagCompound recovery = compound.getCompoundTag(uuid);
-            Class<? extends IPlayerInformation> plClass = TogetherRegistries.getPlayerInformationClass(recovery.getString("PlayerID"));
-            if (plClass != null) {
-                try {
-                    IPlayerInformation info = plClass.newInstance();
-                    info.readFromNBT(recovery.getCompoundTag("ID"));
-                    for (String id : recovery.getKeySet()) {
-                        if (!id.equalsIgnoreCase("ID") && !id.equalsIgnoreCase("PlayerID")) {
-                            offlineRecoveries.put(info, recovery.getCompoundTag(id));
-                        }
-                    }
-                } catch (InstantiationException | IllegalAccessException e) {
-                    e.printStackTrace();
-                }
-            }
         }
     }
 }
